Improved depth first search as the previous one would fail in case of unconnected graphs
